+++ src/editor/core/draw/control/select/SelectControl.ts
@@ -172,7 +172,7 @@ export class SelectControl implements IControlInstance {
       formatElementContext(elementList, [newElement], startIndex, {
         editorOptions: this.options
       })
-      draw.spliceElementList(elementList, start + i, 0, newElement)
+      draw.spliceElementList(elementList, start + i, 0, [newElement])
     }
     return start + data.length - 1
   }
@@ -394,7 +394,7 @@ export class SelectControl implements IControlInstance {
       formatElementContext(elementList, [newElement], prefixIndex, {
         editorOptions: this.options
       })
-      draw.spliceElementList(elementList, start + i, 0, newElement)
+      draw.spliceElementList(elementList, start + i, 0, [newElement])
     }
     // 设置状态
     this.control.setControlProperties(