+++ src/editor/core/event/GlobalEvent.ts
@@ -50,8 +50,7 @@ export class GlobalEvent {
 
   private addEvent() {
     window.addEventListener('blur', this.clearSideEffect)
-    document.addEventListener('keyup', this.setRangeStyle)
-    document.addEventListener('click', this.clearSideEffect)
+    document.addEventListener('mousedown', this.clearSideEffect)
     document.addEventListener('mouseup', this.setCanvasEventAbility)
     document.addEventListener('wheel', this.setPageScale, { passive: false })
     document.addEventListener('visibilitychange', this._handleVisibilityChange)
@@ -60,8 +59,7 @@ export class GlobalEvent {
 
   public removeEvent() {
     window.removeEventListener('blur', this.clearSideEffect)
-    document.removeEventListener('keyup', this.setRangeStyle)
-    document.removeEventListener('click', this.clearSideEffect)
+    document.removeEventListener('mousedown', this.clearSideEffect)
     document.removeEventListener('mouseup', this.setCanvasEventAbility)
     document.removeEventListener('wheel', this.setPageScale)
     document.removeEventListener(
@@ -81,10 +79,7 @@ export class GlobalEvent {
       (node: any) => pageList.includes(node),
       true
     )
-    if (innerEditorDom) {
-      this.setRangeStyle()
-      return
-    }
+    if (innerEditorDom) return
     // 编辑器外部组件dom
     const outerEditorDom = findParent(
       target,
@@ -93,7 +88,6 @@ export class GlobalEvent {
       true
     )
     if (outerEditorDom) {
-      this.setRangeStyle()
       this.watchCursorActive()
       return
     }
@@ -110,10 +104,6 @@ export class GlobalEvent {
   public setCanvasEventAbility = () => {
     this.canvasEvent.setIsAllowDrag(false)
     this.canvasEvent.setIsAllowSelection(false)
-  }
-
-  public setRangeStyle = () => {
-    this.range.setRangeStyle()
   }
 
   public watchCursorActive() {