+++ src/editor/interface/Watermark.ts
@@ -1,3 +1,5 @@
+import { NumberType } from '../dataset/enum/Common'
+
 export interface IWatermark {
   data: string
   color?: string
@@ -5,5 +7,6 @@ export interface IWatermark {
   size?: number
   font?: string
   repeat?: boolean
+  numberType?: NumberType
   gap?: [horizontal: number, vertical: number]
 }