+++ src/editor/utils/element.ts
@@ -9,6 +9,7 @@ import {
   splitText
 } from '.'
 import {
+  BlockType,
   EditorMode,
   ElementType,
   IEditorOption,
@@ -21,6 +22,7 @@ import {
   TableBorder,
   TdBorder
 } from '..'
+import { IFrameBlock } from '../core/draw/particle/block/modules/IFrameBlock'
 import { LaTexParticle } from '../core/draw/particle/latex/LaTexParticle'
 import { NON_BREAKING_SPACE, ZERO } from '../dataset/constant/Common'
 import {
@@ -633,7 +635,38 @@ export function zipElementList(
       continue
     }
     // 优先处理虚拟元素，后表格、超链接、日期、控件特殊处理
-    if (element.titleId && element.level) {
+    if (element.areaId && element.area) {
+      const areaId = element.areaId
+      const area = element.area
+      // 收集并压缩数据
+      const valueList: IElement[] = []
+      while (e < elementList.length) {
+        const areaE = elementList[e]
+        if (areaId !== areaE.areaId) {
+          e--
+          break
+        }
+        delete areaE.area
+        delete areaE.areaId
+        valueList.push(areaE)
+        e++
+      }
+      const areaElementList = zipElementList(valueList, options)
+      // 不归类区域元素
+      if (isClassifyArea) {
+        const areaElement: IElement = {
+          type: ElementType.AREA,
+          value: '',
+          areaId,
+          area
+        }
+        areaElement.valueList = areaElementList
+        element = areaElement
+      } else {
+        zipElementListData.splice(e, 0, ...areaElementList)
+        continue
+      }
+    } else if (element.titleId && element.level) {
       // 标题处理
       const titleId = element.titleId
       if (titleId) {
@@ -687,37 +720,6 @@ export function zipElementList(
         }
         listElement.valueList = zipElementList(valueList, options)
         element = listElement
-      }
-    } else if (element.areaId && element.area) {
-      const areaId = element.areaId
-      const area = element.area
-      // 收集并压缩数据
-      const valueList: IElement[] = []
-      while (e < elementList.length) {
-        const areaE = elementList[e]
-        if (areaId !== areaE.areaId) {
-          e--
-          break
-        }
-        delete areaE.area
-        delete areaE.areaId
-        valueList.push(areaE)
-        e++
-      }
-      const areaElementList = zipElementList(valueList, options)
-      // 不归类区域元素
-      if (isClassifyArea) {
-        const areaElement: IElement = {
-          type: ElementType.AREA,
-          value: '',
-          areaId,
-          area
-        }
-        areaElement.valueList = areaElementList
-        element = areaElement
-      } else {
-        zipElementListData.splice(e, 0, ...areaElementList)
-        continue
       }
     } else if (element.type === ElementType.TABLE) {
       // 分页表格先进行合并
@@ -1249,6 +1251,37 @@ export function createDomFromElementList(
           img.height = element.height!
         }
         clipboardDom.append(img)
+      } else if (element.type === ElementType.BLOCK) {
+        if (element.block?.type === BlockType.VIDEO) {
+          const src = element.block.videoBlock?.src
+          if (src) {
+            const video = document.createElement('video')
+            video.style.display = 'block'
+            video.controls = true
+            video.src = src
+            video.width = element.width! || options?.width || window.innerWidth
+            video.height = element.height!
+            clipboardDom.append(video)
+          }
+        } else if (element.block?.type === BlockType.IFRAME) {
+          const { src, srcdoc } = element.block.iframeBlock || {}
+          if (src || srcdoc) {
+            const iframe = document.createElement('iframe')
+            iframe.sandbox.add(...IFrameBlock.sandbox)
+            iframe.style.display = 'block'
+            iframe.style.border = 'none'
+            if (src) {
+              iframe.src = src
+            } else if (srcdoc) {
+              iframe.srcdoc = srcdoc
+            }
+            iframe.width = `${
+              element.width || options?.width || window.innerWidth
+            }`
+            iframe.height = `${element.height!}`
+            clipboardDom.append(iframe)
+          }
+        }
       } else if (element.type === ElementType.SEPARATOR) {
         const hr = document.createElement('hr')
         clipboardDom.append(hr)
@@ -1481,6 +1514,39 @@ export function getElementListByHTML(
               height,
               value: src,
               type: ElementType.IMAGE
+            })
+          }
+        } else if (node.nodeName === 'VIDEO') {
+          const { src, width, height } = node as HTMLVideoElement
+          if (src && width && height) {
+            elementList.push({
+              value: '',
+              type: ElementType.BLOCK,
+              block: {
+                type: BlockType.VIDEO,
+                videoBlock: {
+                  src
+                }
+              },
+              width,
+              height
+            })
+          }
+        } else if (node.nodeName === 'IFRAME') {
+          const { src, srcdoc, width, height } = node as HTMLIFrameElement
+          if ((src || srcdoc) && width && height) {
+            elementList.push({
+              value: '',
+              type: ElementType.BLOCK,
+              block: {
+                type: BlockType.IFRAME,
+                iframeBlock: {
+                  src,
+                  srcdoc
+                }
+              },
+              width: parseInt(width),
+              height: parseInt(height)
             })
           }
         } else if (node.nodeName === 'TABLE') {